+++ src/Bodoconsult.App/DataExportServices/ByteArrayAndMemoryDataExportService.cs
@@ -1,17 +1,20 @@
 ﻿// Copyright (c) Bodoconsult EDV-Dienstleistungen GmbH.  All rights reserved.
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+namespace Bodoconsult.App.DataExportServices;
 
-namespace Bodoconsult.App.DataExportServices
+/// <summary>
+/// Data export service for byte arrays
+/// </summary>
+public class ByteArrayDataExportService : DataExportServiceBase<byte[]>
 {
-
     /// <summary>
-    /// Data export service for byte arrays
+    /// Converts an object of type T into a ReadOnlyMemory&lt;byte&gt; instance
     /// </summary>
-    public class ByteArrayDataExportService : DataExportServiceBase<byte[]>
-    { }
+    /// <param name="data">Data object to serialize</param>
+    /// <returns>Byte array</returns>
+    /// <exception cref="NotSupportedException">Thrown if type T is NOT string, ReadOnlyMemory&lt;byte&gt; or byte[]</exception>
+    public override ReadOnlyMemory<byte> ToMemory(byte[] data)
+    {
+        return data;
+    }
 }