+++ src/Bodoconsult.App/DataExportServices/JsonDataExportService.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Bodoconsult EDV-Dienstleistungen GmbH.  All rights reserved.
 
 using System.Text;
+using System.Text.Json;
 
 namespace Bodoconsult.App.DataExportServices;
 
@@ -9,11 +10,20 @@ namespace Bodoconsult.App.DataExportServices;
 /// </summary>
 public class JsonDataExportService<T> : DataExportServiceBase<T> where T : class
 {
-
     /// <summary>
     /// Default ctor
     /// </summary>
     public JsonDataExportService()
+    {
+        HeaderData = Encoding.GetBytes("[");
+        FooterData = Encoding.GetBytes("]");
+        TokenSeparatorData = Encoding.GetBytes(",");
+    }
+
+    /// <summary>
+    /// Ctor supplying an encoding
+    /// </summary>
+    public JsonDataExportService(Encoding encoding) : base(encoding)
     {
         HeaderData = Encoding.GetBytes("[");
         FooterData = Encoding.GetBytes("]");
@@ -27,7 +37,7 @@ public class JsonDataExportService<T> : DataExportServiceBase<T> where T : class
     /// <returns>Byte array</returns>
     public override ReadOnlyMemory<byte> ToMemory(T data)
     {
-        var s = System.Text.Json.JsonSerializer.Serialize(data);
+        var s = JsonSerializer.Serialize(data);
         var b = Encoding.GetBytes(s);
         return b;
     }