+++ src/Bodoconsult.App/DataExportServices/StringDataExportService.cs
@@ -6,4 +6,18 @@ namespace Bodoconsult.App.DataExportServices;
 /// Data export service for UTF8 strings
 /// </summary>
 public class StringDataExportService : DataExportServiceBase<string>
-{ }
+{
+    /// <summary>
+    /// Converts an object of type T into a ReadOnlyMemory&lt;byte&gt; instance
+    /// </summary>
+    /// <param name="data">Data object to serialize</param>
+    /// <returns>Byte array</returns>
+    /// <exception cref="NotSupportedException">Thrown if type T is NOT string, ReadOnlyMemory&lt;byte&gt; or byte[]</exception>
+    public override ReadOnlyMemory<byte> ToMemory(string data)
+    {
+        data ??= string.Empty;
+
+        var arr = Encoding.GetBytes(data);
+        return arr;
+    }
+}