+++ tests/Bodoconsult.App.Test/BufferPool/BufferPoolResetableTests.cs
@@ -1,11 +1,11 @@
-﻿// Copyright (c) Bodoconsult EDV-Dienstleistungen GmbH. All rights reserved.
+﻿// Copyright (c) Bodoconsult EDV-Dienstleistungen GmbH.  All rights reserved.
 
 using Bodoconsult.App.BufferPool;
 
 namespace Bodoconsult.App.Test.BufferPool;
 
 [TestFixture]
-internal class BufferPoolTests
+internal class BufferPoolResetableTests
 {
     private const int NumberOfItems = 1000;
 
@@ -13,7 +13,7 @@ internal class BufferPoolTests
     public void Allocate_ValidSetup_QueueFilled()
     {
         // Arrange 
-        var myPool = new BufferPool<byte[]>(() => new byte[65535]);
+        var myPool = new BufferPoolResetable<TestDataResetable>(() => new TestDataResetable());
 
         // Act  
         myPool.Allocate(NumberOfItems);
@@ -26,7 +26,7 @@ internal class BufferPoolTests
     public void Dequeue_ValidSetup_InstanceDequeued()
     {
         // Arrange 
-        var myPool = new BufferPool<byte[]>(() => new byte[65535]);
+        var myPool = new BufferPoolResetable<TestDataResetable>(() => new TestDataResetable());
         myPool.Allocate(NumberOfItems);
 
         // Act  
@@ -42,7 +42,7 @@ internal class BufferPoolTests
     public void Enqueue_ValidSetup_InstanceEnqueued()
     {
         // Arrange 
-        var myPool = new BufferPool<byte[]>(() => new byte[65535]);
+        var myPool = new BufferPoolResetable<TestDataResetable>(() => new TestDataResetable());
         myPool.Allocate(1000);
 
         var buffer = myPool.Dequeue();