+++ tests/Bodoconsult.App.Test/DataExportServices/StringDataExportServiceTests.cs
@@ -58,21 +58,6 @@ internal class StringDataExportServiceTests
     }
 
     [Test]
-    public void Start_ValidDefaultSetup_CurrentFilePathSet()
-    {
-        // Arrange 
-        var service = new StringDataExportService();
-
-        // Act  
-        service.Start();
-
-        service.Stop();
-
-        // Assert
-        Assert.That(string.IsNullOrEmpty(service.CurrentFilePath), Is.False);
-    }
-
-    [Test]
     public void Add_ValidDefaultSetup_FileWritten()
     {
         // Arrange